Do not allow rescheduling of cancelled appointments
Cancelling an appointment creates a new slot for the same time - the idea
being that the original slot is now redundant. But allowing a cancelled
appointment to be rescheduled makes the slot available for use again.

On 19th Oct 2018 we received a report that a particular room for a
particular time had 3 slots with 1 cancelled appointment and 2 pending
appointments.

The events that led up to that were:

```
At the start.. just 1 slot existed..

1. Appointment 3091 (Ann) for 7/11 cancelled at 12:03 on 15/10
2. Automatically creates a new slot 9086 for 7/11

Currently 2 slots.. (1 available and 1 with a cancelled appointment)

3. Appointment 3502 (Wai) rescheduled to slot 9086 at 12:04 on 15/10

Currently 2 slots.. (1 with a pending appointment and 1 with a cancelled appointment)

4. Appointment 3091 (Ann) rescheduled to slot 2823 for 8/11 at 12:04 on 15/10

Currently 2 slots.. (1 available and 1 with a pending appointment)

5. Appointment 3556 (Rebecca) created to slot 2826 at 18:01 on 15/10

Currently 2 slots.. (2 with pending appointments)

6. Appointment 3556 (Rebecca) cancelled at 14:15:34 on 17/10
7. Automatically creates a new slot 9093 for 7/11

Currently 3 slots.. (1 with cancelled appointment, 1 with pending, 1 available)

8. Appointment 3591 (Nicola) created for slot 9093 at 17:04 on 17/10

Currently 3 slots.. (1 with cancelled appointment, 2 with pending appointments)
```

If 4. hadn't happened, then it wouldn't have been possible to end up
with 2 slots with 2 pending appointments.